makefile: simplify building of the docker image
Remove the hack/build_docker_image.sh script. The majority of the args presented
to Docker are already part of our Makefile, so re-use them.

Also, `generate-distribution' depends on the `builder' tool only, which should
speed things up, since we don't have to pull all tools for generating the
distro.

Also, bump the image to golang:1.24