refactor: simplify building of collector and docker image (#7)
* makefile: simplify building of otel collector

Remove the hack/build_distribution.sh script. We already have the required vars
in the Makefile, so re-use them.

* makefile: simplify COMPONENT_DIRS var

EXCL_TOOLS_DIR and EXCL_BUILD_DIR are actually find(1) patterns and are
confusing when they stand on their own, so add them to the actual find(1)
command.

Also remove the BUILD_ARCH var, since we don't use it.

* go: bump min Go version to 1.24

* makefile: simplify building of the docker image

Remove the hack/build_docker_image.sh script. The majority of the args presented
to Docker are already part of our Makefile, so re-use them.

Also, `generate-distribution' depends on the `builder' tool only, which should
speed things up, since we don't have to pull all tools for generating the
distro.

Also, bump the image to golang:1.24

* makefile: add targets to print revision and effective version

* github/workflows/build: pass version info as build args

* makefile: use existing REVISION var as build arg

* makefile: correctly pass BUILD_DATE build arg

* github/workflows: rename job to `get-version-info`